+++ components/src/Select/index.js
@@ -1,2 +1,1 @@
 export { default as Select } from "./Select";
-export { default as ReactSelect } from "./ReactSelect";
